pyproject.toml
[tool.poetry]
name = "project-api-rest-django"
version = "0.1.0"
description = ""
authors = ["Jose Felipe <felipe.brx.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
python-decouple = "^3.8"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
pillow = "^10.2.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
